+++ CHANGELOG.md
@@ -11,8 +11,11 @@
 
 ### Additions
 
-- Added parameter for showing grid when plotting diel cycles (`diive.core.plotting.dielcycle.DielCycle.plot`)
-- Added parameter for hiding x-ticklabels when plotting diel cycles (`diive.core.plotting.dielcycle.DielCycle.plot`)
+- Added parameters when plotting diel cycles:
+    - Added parameter `show_xticklabels` for showing grid
+    - Added parameter `show_xlabel` for showing x-ticklabels
+    - Added parameter `show_legend` for showing legend
+    - (`diive.core.plotting.dielcycle.DielCycle.plot`)
 
 ## v0.87.0 | 17 May 2025